requirements.txt
asgiref==3.9.2
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
distlib==0.4.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
google-auth==2.41.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.4.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
rest-framework-simplejwt==0.0.2
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
whitenoise==6.11.0
